# openapi.json
{
  "components": {
    "schemas": {
...
      "ChatIn": {
        "properties": {
          "grounded": {
            "default": true,
            "title": "Grounded",
            "type": "boolean"
          },
          "learning_event": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LearningEventIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageIn"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Messages",
            "type": "array"
          },
          "mode": {
            "default": "explain",
            "enum": [
              "explain",
              "socratic",
              "examiner",
              "study_partner",
              "feynman"
            ],
            "title": "Mode",
            "type": "string"
          },
          "notebook_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notebook Id"
          },
          "session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "required": [
          "messages"
        ],
        "title": "ChatIn",
        "type": "object"
      },
...
      "CredentialCreate": {
        "description": "Write-only. There is deliberately no response model containing ``api_key``.",
        "properties": {
          "api_key": {
            "maxLength": 500,
            "minLength": 8,
            "title": "Api Key",
            "type": "string"
          },
          "label": {
            "default": "default",
            "maxLength": 120,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "provider": {
            "maxLength": 50,
            "minLength": 2,
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "api_key"
        ],
        "title": "CredentialCreate",
        "type": "object"
      },
      "CredentialOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "last4": {
            "title": "Last4",
            "type": "string"
          },
          "last_used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "last_verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Verified At"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "verification_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Error"
          }
        },
        "required": [
          "id",
          "provider",
          "label",
          "last4",
          "created_at",
          "last_used_at",
          "last_verified_at",
          "verification_error"
        ],
        "title": "CredentialOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TeachingSessionOut": {
        "description": "A lesson as the client resumes it: where it is, and what was said.\n\nThe professor's structured decisions and metadata stay server-side; this is\nthe learner-facing shape — enough to show \"where we stopped\" and to replay\nthe transcript, nothing the interface would have to hide.",
        "properties": {
          "current_concept": {
            "title": "Current Concept",
            "type": "string"
          },
          "current_topic": {
            "title": "Current Topic",
            "type": "string"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "journey_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Journey Id"
          },
          "last_turn_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Turn At"
          },
          "learning_goal": {
            "title": "Learning Goal",
            "type": "string"
          },
          "notebook_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notebook Id"
          },
          "plan": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Plan",
            "type": "array"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "turn_count": {
            "title": "Turn Count",
            "type": "integer"
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/TeachingTurnOut"
            },
            "title": "Turns",
            "type": "array"
          }
        },
        "required": [
          "id",
          "notebook_id",
          "learning_goal",
          "subject",
          "current_topic",
          "current_concept",
          "plan",
          "turn_count",
          "last_turn_at",
          "ended_at",
          "turns"
        ],
        "title": "TeachingSessionOut",
        "type": "object"
      },
...
      "UsageOut": {
        "properties": {
          "completion_tokens": {
            "title": "Completion Tokens",
            "type": "integer"
          },
          "cost_cents": {
            "title": "Cost Cents",
            "type": "number"
          },
          "prompt_tokens": {
            "title": "Prompt Tokens",
            "type": "integer"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "task": {
            "title": "Task",
            "type": "string"
          }
        },
        "required": [
          "task",
          "provider",
          "prompt_tokens",
          "completion_tokens",
          "cost_cents"
        ],
        "title": "UsageOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "NOEMA — open-source adaptive learning platform.\n\nCookie sessions authenticate the web app; mutations require the CSRF token returned\nat login in the `x-csrf-token` header.",
    "title": "NOEMA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ai/chat": {
      "post": {
        "description": "Stream a tutor reply as Server-Sent Events.\n\nWith a notebook, the reply is grounded: chunks are retrieved, numbered into the\nprompt, and every citation is validated on the way out. Without one, it is an\nordinary tutor conversation.",
        "operationId": "chat_api_v1_ai_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/v1/ai/credentials": {
      "get": {
        "operationId": "list_credentials_api_v1_ai_credentials_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CredentialOut"
                  },
                  "title": "Response List Credentials Api V1 Ai Credentials Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Credentials",
        "tags": [
          "ai"
        ]
      },
      "post": {
        "operationId": "create_credential_api_v1_ai_credentials_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Credential",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/v1/ai/credentials/{credential_id}": {
      "delete": {
        "operationId": "delete_credential_api_v1_ai_credentials__credential_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "credential_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Credential Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Credential",
        "tags": [
          "ai"
        ]
      }
    },
...
    "/api/v1/ai/professor": {
      "post": {
        "description": "Stream one turn of the Professor (V3).\n\nAdditive to ``chat()`` above — ``POST /ai/chat`` stays the direct,\nmanual-``mode`` path. Here the server decides: it finds or starts the\nlearning journey (goal → curriculum), records the learning event the\nclient sent, chooses the move before any model speaks, assembles the\nprompt from *stored* state under a token budget, and streams the reply\nas prose plus structured events (``block``, ``mino``, ``flashcards``,\n``checkpoint``, ``mastery``, ``memory``). See ``noema/professor/``.",
        "operationId": "professor_chat_api_v1_ai_professor_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Professor Chat",
        "tags": [
          "ai"
        ]
      }
    },
...
    "/api/v1/ai/sessions/latest": {
      "get": {
        "description": "Where the learner left off — the home screen's \"Continue learning\".",
        "operationId": "latest_session_api_v1_ai_sessions_latest_get",
        "parameters": [
          {
            "in": "query",
            "name": "notebook_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Notebook Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/TeachingSessionOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Latest Session Api V1 Ai Sessions Latest Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Latest Session",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/v1/ai/sessions/{session_id}": {
      "get": {
        "description": "A lesson, resumed: its state and transcript, oldest turn first.",
        "operationId": "get_session_api_v1_ai_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeachingSessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/v1/ai/usage": {
      "get": {
        "operationId": "usage_api_v1_ai_usage_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UsageOut"
                  },
                  "title": "Response Usage Api V1 Ai Usage Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Usage",
        "tags": [
          "ai"
        ]
      }
    },
...
  }
}